+++ app/src/main/java/net/aiscope/gdd_app/repository/SampleRepository.kt
@@ -10,8 +10,6 @@ interface SampleRepository {
 
     fun all(): List<Sample>
     suspend fun lastSaved(): Sample?
-
-    fun storeCollection(sample: Sample)
 }
 
 class SampleRepositoryException(message: String): Exception(message)
