+++ app/src/test/java/net/aiscope/gdd_app/presentation/SampleCompletionViewModelTest.kt
@@ -20,6 +20,7 @@ import net.aiscope.gdd_app.model.WaterType
 import net.aiscope.gdd_app.network.RemoteStorage
 import net.aiscope.gdd_app.repository.MicroscopistRepository
 import net.aiscope.gdd_app.repository.SampleRepository
+import net.aiscope.gdd_app.repository.SampleRepositoryFirestore
 import net.aiscope.gdd_app.ui.sample_completion.SampleCompletionViewModel
 import org.junit.Assert.assertEquals
 import org.junit.Before
@@ -75,6 +76,9 @@ class SampleCompletionViewModelTest {
     @Mock
     private lateinit var microscopistRepository: MicroscopistRepository
 
+    @Mock
+    private lateinit var sampleRepositoryFirestore: SampleRepositoryFirestore
+
     @Before
     fun before() {
         coroutinesTestRule.runBlockingTest {
@@ -95,7 +99,10 @@ class SampleCompletionViewModelTest {
 
             whenever(repository.current()).thenReturn(sample)
 
-            viewModel = SampleCompletionViewModel(repository, remoteStorage, context, microscopistRepository)
+            viewModel = SampleCompletionViewModel(
+                repository, remoteStorage, context,
+                microscopistRepository, sampleRepositoryFirestore
+            )
         }
     }
